// frontend/packages/web/src/enums/RawFeature.ts
export enum RawFeature {
  ANAG_SUBJECT_BASE = '121',
  ANAG_SUBJECT_CATEGORY = '131',
  ANAG_SUBJECT_ORGUNIT = '126',
  ASST_ESTATE_BASE = '211',
  ASST_ESTATE_CATALOGUE = '212',
  ASST_ESTATEUNIT_BASE = '214',
  ASST_ESTATEUNIT_VARIATION = '215',
  ASST_CADASTRALUNIT_BASE = '217',
  ASST_CATALOGUE_CONFIG = '230',
  ASST_CADASTRAL_LAND_CATEGORY = '240',
  ASST_ASSET_TAX_BASE = '250',
  COMMON_VATRATES = '311',
  COMMON_INTERESTRATES = '321',
  COMMON_ACCOUNTINGITEMS = '322',
  COMMON_BILLITEMTYPES = '323',
  COMMON_REVALUATIONDATA = '324',
  COMMON_CITIES = '331',
  COMMON_TAX_RATES = '341',
  COMMON_REPORT_GENERATORS = '351',
  DOCUMENTS_BASE = '401',
  PROP_CONTRACT_BASE = '501',
  PROP_CONTRACT_TYPES = '511',
  PROP_REGISTRATION_OFFICES = '512',
  PROP_BILL_BASE = '521',
  PROP_ADMINISTRATION_BASE = '531',
  PROP_REGISTRY_COMMUNICATION = '541',
  PROP_REGISTRATION_PAYMENT = '551',
  NRGY_TYPE_BASE = '601',
  NRGY_SERVICE_BASE = '611',
  NRGY_COSTCHARGE_BASE = '621',
  ECON_TAX_CREDIT_BASE = '701',
  FCLT_ESTATE_UNIT_GROUP_BASE = '800',
  FCLT_CONTRACT_BASE = '810',
  FCLT_CONTRACT_TEMPLATES = '811',
  FCLT_CONTRACT_TYPES = '812',
  FCLT_SLA_BASE = '820',
  FCLT_CONFIG = '830',
  FCLT_PENALTY_BASE = '840',
  FCLT_PRICE_LIST_BASE = '850',
  FCLT_PRICE_LIST_ARTICLES = '851',
  FCLT_TICKET = '860',
  FCLT_SERVICE_CATEGORY = '865',
  FCLT_SERVICE = '870',
  ADMIN_SETTINGS = '911',
  ADMIN_USERS_AND_GROUPS = '921',
  ADMIN_AUDIT_LOG = '922',
  ADMIN_CONFIG = '923',
}